services/ecosense-sync/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.6.0
pydantic-settings==2.1.0
